Move the filtering of sensitive data to the config class
